Changed the file format on disk
This removes the null gap between the nonce and the cyphertext. This ensures
that we do not need to worry about sequential nulls in the nonce but, it
depends on aead.NonceLength() not ever changing. Due to this, I've added a file
encrypted with this version that should never fail to be decrypted.